backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
chromadb
google-generativeai
python-dotenv
pillow
numpy
scikit-learn
python-multipart